# backend/static/docs/swagger.yaml
basePath: /api/
definitions:
  db.Post:
    properties:
      body:
        type: string
      created_at:
        type: string
      deletedAt:
        $ref: "#/definitions/gorm.DeletedAt"
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: "#/definitions/db.User"
      user_id:
        type: integer
    type: object
  db.User:
    properties:
      created_at:
        type: string
      deletedAt:
        $ref: "#/definitions/gorm.DeletedAt"
      email:
        type: string
      id:
        description: gorm.Model `json:"-"`
        type: integer
      name:
        type: string
      posts:
        description: Relationships
        items:
          $ref: "#/definitions/db.Post"
        type: array
      updated_at:
        type: string
      username:
        type: string
      verification_code:
        type: string
      verification_code_expiry:
        type: string
      verified:
        description: Verification
        type: boolean
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.CreatePost:
    properties:
      body:
        type: string
      title:
        type: string
      user:
        $ref: "#/definitions/db.User"
    type: object
...
  models.PaginatedPosts:
    properties:
      data:
        items:
          $ref: "#/definitions/db.Post"
        type: array
      meta: {}
    type: object
...
  models.UpdatePost:
    properties:
      body:
        type: string
      title:
        type: string
      user:
        $ref: "#/definitions/db.User"
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a template API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Template
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
        - application/json
      description:
        Accepts limit, offset, order, search (added as non-case sensitive
        LIKE), title, body as query parameters
      parameters:
        - description: limit
          in: query
          name: limit
          required: true
          type: integer
        - description: offset
          in: query
          name: offset
          type: integer
        - description: "order by eg. (asc) "
          in: query
          name: order
          type: integer
        - description: search (added to all string conditions as LIKE SQL search)
          in: query
          name: search
          type: string
        - description: title
          in: query
          name: title
          type: string
        - description: body
          in: query
          name: body
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.PaginatedPosts"
        "400":
          description: Error extracting query params
          schema:
            type: string
      security:
        - BearerToken: []
      summary: Finds a list of posts
      tags:
        - Post
    post:
      consumes:
        - application/json
      description: Creates a new post
      parameters:
        - description: New Post
          in: body
          name: post
          required: true
          schema:
            $ref: "#/definitions/models.CreatePost"
      produces:
        - text/plain
      responses:
        "201":
          description: Post creation successful!
          schema:
            type: string
        "400":
          description: Post creation failed.
          schema:
            type: string
      security:
        - BearerToken: []
      summary: Create Post
      tags:
        - Post
  /posts/{id}:
    delete:
      consumes:
        - application/json
      description: Deletes an existing post
      parameters:
        - description: Post ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Deletion successful!
          schema:
            type: string
        "400":
          description: Failed post deletion
          schema:
            type: string
      security:
        - BearerToken: []
      summary: Delete Post
      tags:
        - Post
    get:
      consumes:
        - application/json
      description: Find a post by ID
      parameters:
        - description: Post ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/db.Post"
        "400":
          description: "Can't find post with ID: {id}"
          schema:
            type: string
      security:
        - BearerToken: []
      summary: Find post
      tags:
        - Post
    put:
      consumes:
        - application/json
      description: Updates an existing post
      parameters:
        - description: Update Post
          in: body
          name: post
          required: true
          schema:
            $ref: "#/definitions/models.UpdatePost"
        - description: Post ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/db.Post"
        "400":
          description: Failed post update
          schema:
            type: string
        "403":
          description: Authentication Token not detected
          schema:
            type: string
      security:
        - BearerToken: []
      summary: Update Post
      tags:
        - Post
...
securityDefinitions:
  BearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"